pyproject.toml
[project]
name = "AgentCrew"
version = "0.3.5"
requires-python = ">=3.12"
classifiers = ["Private :: Do not upload"]
description = "Multi-Agents Interactive Chat Tool"
dependencies = [
  "firecrawl",
  "click",
  "python-dotenv",
  "anthropic",
  "pytest",
  "prompt-toolkit>=3.0.50",
  "rich>=13.9.4",
  "pyperclip>=1.9.0",
  "tavily-python>=0.5.1",
  "pillow>=11.1.0",
  "groq>=0.18.0",
  "chromadb>=0.6.3",
  "openai>=1.65.2",
  "tree-sitter>=0.23.2",
  "mcp>=1.3.0",
  "docling>=2.26.0",
  "google-genai>=1.7.0",
  "toml>=0.10.2",
  "pyside6>=6.8.3",
  "markdown>=3.7",
  "tree-sitter-language-pack>=0.7.0",
  "a2a-samples",
  "mem0ai>=0.1.95",
]


[tool.uv.sources]
a2a-samples = { git = "https://github.com/google/A2A.git", subdirectory = "samples/python" }

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.scripts]
AgentCrew = "AgentCrew.main:cli_prod"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["./"]

[dependency-groups]
dev = ["pygments>=2.19.1"]
